+++ src/main/java/uet/model/ItemFactory/ElectronicsFactory.java
@@ -3,9 +3,11 @@ package uet.model.ItemFactory;
 import uet.model.items.Electronics;
 import uet.model.items.Item;
 
+import java.math.BigDecimal;
+
 public class ElectronicsFactory extends ItemFactory {
     @Override
-    public Item createItem(String name, String description, double startingPrice, String imageUrl, String... extraInfo) {
+    public Item createItem(String name, String description, BigDecimal startingPrice, String imageUrl, String... extraInfo) {
         String brand = extraInfo[0];
         int warrantyMonths = Integer.parseInt(extraInfo[1]);
         return new Electronics(name,description,startingPrice,imageUrl,brand,warrantyMonths);
