+++ src/main/java/uet/model/items/Electronics.java
@@ -1,11 +1,14 @@
 package uet.model.items;
+
+import java.math.BigDecimal;
+
 public class Electronics extends Item{
     private String brand;
     private int warrantyMonths;     // tháng bảo
     public Electronics() {
         super();
     }
-    public Electronics(String name, String description, double startingPrice, String imageUrl, String brand, int warrantyMonths) {
+    public Electronics(String name, String description, BigDecimal startingPrice, String imageUrl, String brand, int warrantyMonths) {
         super(name, description, startingPrice, imageUrl);
         this.brand = brand;
         this.warrantyMonths = warrantyMonths;
